Revert "feat(a2a): SSE streaming + external-IdP/OAuth auth, and a unified Cre…"
This reverts commit 814951117739e44863facf48d56b0c723ddd6730.
