+++ platform/backend/src/routes/mcp-oauth-clients.test.ts
@@ -130,7 +130,7 @@ describe("mcpOauthClientsRoutes", () => {
     ).toEqual(["Searchable Service"]);
   });
 
-  test("rejects an llm_proxy agent as an allowed gateway", async ({
+  test("rejects a non-gateway agent as an allowed gateway", async ({
     makeAgent,
   }) => {
     const llmProxy = await makeAgent({
@@ -149,30 +149,7 @@ describe("mcpOauthClientsRoutes", () => {
     });
 
     expect(response.statusCode).toBe(404);
-    expect(response.json().error.message).toContain(
-      "MCP gateway or agent not found",
-    );
-  });
-
-  test("accepts an A2A agent (agentType='agent') as an allowed gateway", async ({
-    makeInternalAgent,
-  }) => {
-    const a2aAgent = await makeInternalAgent({
-      organizationId,
-      name: "A2A Agent",
-    });
-
-    const response = await app.inject({
-      method: "POST",
-      url: "/api/mcp-oauth-clients",
-      payload: {
-        name: "A2A OAuth Client",
-        allowedGatewayIds: [a2aAgent.id],
-      },
-    });
-
-    expect(response.statusCode).toBe(200);
-    expect(response.json().allowedGatewayIds).toContain(a2aAgent.id);
+    expect(response.json().error.message).toContain("MCP gateway not found");
   });
 
   test("rejects a gateway from another organization", async ({
@@ -196,9 +173,7 @@ describe("mcpOauthClientsRoutes", () => {
     });
 
     expect(response.statusCode).toBe(404);
-    expect(response.json().error.message).toContain(
-      "MCP gateway or agent not found",
-    );
+    expect(response.json().error.message).toContain("MCP gateway not found");
   });
 
   test("creates an authorization_code client registered as a confidential, PKCE client", async () => {
@@ -369,8 +344,6 @@ describe("mcpOauthClientsRoutes", () => {
     });
 
     expect(response.statusCode).toBe(404);
-    expect(response.json().error.message).toContain(
-      "MCP gateway or agent not found",
-    );
+    expect(response.json().error.message).toContain("MCP gateway not found");
   });
 });
