+++ platform/frontend/src/app/_parts/sidebar.tsx
@@ -20,7 +20,6 @@ import {
   FolderKanban,
   Github,
   Inbox,
-  KeyRound,
   type LucideIcon,
   MessageCircle,
   MessagesSquare,
@@ -282,6 +281,14 @@ const contentNavGroups: NavGroup[] = [
         icon: Waypoints,
         customIsActive: (pathname: string) =>
           pathname.startsWith("/mcp/gateways"),
+        subItems: [
+          {
+            title: "Credentials",
+            url: "/mcp/credentials/oauth-clients",
+            customIsActive: (pathname: string) =>
+              pathname.startsWith("/mcp/credentials"),
+          },
+        ],
       },
     ],
   },
@@ -293,6 +300,14 @@ const contentNavGroups: NavGroup[] = [
         url: "/llm/proxies",
         icon: Network,
         customIsActive: (pathname: string) => pathname === "/llm/proxies",
+        subItems: [
+          {
+            title: "Credentials",
+            url: "/llm/credentials/virtual-keys",
+            customIsActive: (pathname: string) =>
+              pathname.startsWith("/llm/credentials"),
+          },
+        ],
       },
       {
         title: "Model Providers",
@@ -302,13 +317,6 @@ const contentNavGroups: NavGroup[] = [
           pathname.startsWith("/llm/model-providers") ||
           pathname.startsWith("/llm/models"),
         dotKey: "nav:model-providers",
-      },
-      {
-        title: "Credentials",
-        url: "/credentials/oauth-clients",
-        icon: KeyRound,
-        customIsActive: (pathname: string) =>
-          pathname.startsWith("/credentials"),
       },
       {
         title: "Costs & Limits",
