+++ platform/frontend/src/components/conversation-search-palette.tsx
@@ -154,9 +154,8 @@ const navigationItems = [
     icon: Key,
     label: "Credentials",
     value: "credentials",
-    keywords:
-      "virtual keys oauth clients client credentials llm mcp gateways agents a2a",
-    href: "/credentials/oauth-clients",
+    keywords: "virtual keys oauth clients client credentials llm",
+    href: "/llm/credentials/virtual-keys",
   },
   {
     icon: MessagesSquare,
